make JSX being its own JSX component
