+++ components/x-topic-search/src/SuggestionList.jsx
@@ -5,18 +5,25 @@ import classNames from 'classnames';
 export default ({ suggestions, searchTerm }) => {
 	 // TODO use x-follow-button
 	const listResults = suggestions.map((suggestion, index) => (
-		<li className={ classNames(styles["suggestion"]) }
-			key={ index }
-			data-trackable="myft-topic"
-			data-concept-id={ suggestion.conceptId }
-			data-trackable-meta={ '{"search-term":"' + searchTerm + '"}' }>
 
-			<a data-trackable="topic-link" className={ classNames(styles["suggestion__name"]) } href={ suggestion.url }>{ suggestion.prefLabel }</a>
+			<li className={ classNames(styles["suggestion"]) }
+				key={ index }
+				data-trackable="myft-topic"
+				data-concept-id={ suggestion.conceptId }
+				data-trackable-meta={ '{"search-term":"' + searchTerm + '"}' }>
 
-			<button>add to myFT</button>
+				<a data-trackable="topic-link" className={ classNames(styles["suggestion__name"]) } href={ suggestion.url }>
+					{ suggestion.prefLabel }
+				</a>
 
-		</li>
+				<button>add to myFT</button>
+
+			</li>
+
 	))
 
-	return listResults;
+	return (
+		<ul className={ classNames(styles["suggestions"]) } aria-live="polite">
+			{ listResults }
+		</ul>);
 };
